Mathlab files from graduate project uploaded
There are two mathlab files. The purpose of the first file, task 1, is to understand how to train data and perform fitting. The second file, task 2, is to apply the learned concepts in order to predict the price of a house, given a k number of attributes, and utilizing different order models.